feat: change the `userId` in `bookmark` table to Long type, remove the `userName` field from the table and refactor the related methods

+++ src/main/java/com/github/learndifferent/mtm/dto/NotificationDTO.java
@@ -102,7 +102,7 @@ public class NotificationDTO implements Serializable {
      * this can be used to store the ID of the user who get the message.
      * </p>
      */
-    private Integer recipientUserId;
+    private Long recipientUserId;
 
     /**
      * ID of the comment
